Update: TransactionFullDto class

+++ src/System/View/Accounting/Transactions/TransactionFullDto.cs
@@ -6,7 +6,7 @@ public class TransactionFullDto
     public string studentId { get; set; }
     public string studentName { get; set; }
     public string enrollmentLabel { get; set; }
-    public int total { get; set; }
+    public decimal total { get; set; }
     public string dateTime { get; set; }
     public int type { get; set; }
     public bool isValid { get; set; }
